ApisDesplegadas/Whisper_AI/requirements.txt
fastapi
uvicorn[standard]
openai-whisper
torch
python-multipart
